Updated integration test project name
