+++ dotnet/src/SemanticKernel.IntegrationTests/AI/OpenAICompletionTests.cs
@@ -5,15 +5,15 @@ using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
 using System.Threading.Tasks;
-using IntegrationTests.TestSettings;
 using Microsoft.Extensions.Configuration;
 using Microsoft.SemanticKernel;
 using Microsoft.SemanticKernel.KernelExtensions;
 using Microsoft.SemanticKernel.Orchestration;
+using SemanticKernel.IntegrationTests.TestSettings;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace IntegrationTests.AI;
+namespace SemanticKernel.IntegrationTests.AI;
 
 public sealed class OpenAICompletionTests : IDisposable
 {
