+++ dotnet/src/SemanticKernel.IntegrationTests/SemanticKernel.IntegrationTests.csproj
@@ -1,8 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <AssemblyName>IntegrationTests</AssemblyName>
-    <RootNamespace>IntegrationTests</RootNamespace>
+    <AssemblyName>SemanticKernel.IntegrationTests</AssemblyName>
+    <RootNamespace>SemanticKernel.IntegrationTests</RootNamespace>
     <TargetFramework>net6.0</TargetFramework>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
