+++ dotnet/src/SemanticKernel.IntegrationTests/TestSettings/AzureOpenAIConfiguration.cs
@@ -2,20 +2,25 @@
 
 using System.Diagnostics.CodeAnalysis;
 
-namespace IntegrationTests.TestSettings;
+namespace SemanticKernel.IntegrationTests.TestSettings;
 
 [SuppressMessage("Performance", "CA1812:Internal class that is apparently never instantiated",
     Justification = "Configuration classes are instantiated through IConfiguration.")]
-internal sealed class OpenAIConfiguration
+internal sealed class AzureOpenAIConfiguration
 {
     public string Label { get; set; }
-    public string ModelId { get; set; }
+
+    public string DeploymentName { get; set; }
+
+    public string Endpoint { get; set; }
+
     public string ApiKey { get; set; }
 
-    public OpenAIConfiguration(string label, string modelId, string apiKey)
+    public AzureOpenAIConfiguration(string label, string deploymentName, string endpoint, string apiKey)
     {
         this.Label = label;
-        this.ModelId = modelId;
+        this.DeploymentName = deploymentName;
+        this.Endpoint = endpoint;
         this.ApiKey = apiKey;
     }
 }
