+++ dotnet/src/SemanticKernel.IntegrationTests/TestSettings/OpenAIConfiguration.cs
@@ -2,25 +2,20 @@
 
 using System.Diagnostics.CodeAnalysis;
 
-namespace IntegrationTests.TestSettings;
+namespace SemanticKernel.IntegrationTests.TestSettings;
 
 [SuppressMessage("Performance", "CA1812:Internal class that is apparently never instantiated",
     Justification = "Configuration classes are instantiated through IConfiguration.")]
-internal sealed class AzureOpenAIConfiguration
+internal sealed class OpenAIConfiguration
 {
     public string Label { get; set; }
-
-    public string DeploymentName { get; set; }
-
-    public string Endpoint { get; set; }
-
+    public string ModelId { get; set; }
     public string ApiKey { get; set; }
 
-    public AzureOpenAIConfiguration(string label, string deploymentName, string endpoint, string apiKey)
+    public OpenAIConfiguration(string label, string modelId, string apiKey)
     {
         this.Label = label;
-        this.DeploymentName = deploymentName;
-        this.Endpoint = endpoint;
+        this.ModelId = modelId;
         this.ApiKey = apiKey;
     }
 }
