+++ dotnet/src/SemanticKernel.IntegrationTests/testsettings.json
